requirements.txt
streamlit
langchain
openai
chromadb
beautifulsoup4
requests
python-dotenv
